+++ src/Silex/Provider/DoctrineServiceProvider.php
@@ -11,8 +11,9 @@
 
 namespace Silex\Provider;
 
+use Pimple\Container;
+use Pimple\ServiceProviderInterface;
 use Silex\Application;
-use Silex\ServiceProviderInterface;
 use Doctrine\DBAL\DriverManager;
 use Doctrine\DBAL\Configuration;
 use Doctrine\Common\EventManager;
@@ -25,7 +26,7 @@ use Symfony\Bridge\Doctrine\Logger\DbalLogger;
  */
 class DoctrineServiceProvider implements ServiceProviderInterface
 {
-    public function register(Application $app)
+    public function register(Container $app)
     {
         $app['db.default_options'] = array(
             'driver'   => 'pdo_mysql',
@@ -63,7 +64,7 @@ class DoctrineServiceProvider implements ServiceProviderInterface
         $app['dbs'] = $app->share(function ($app) {
             $app['dbs.options.initializer']();
 
-            $dbs = new \Pimple();
+            $dbs = new Container();
             foreach ($app['dbs.options'] as $name => $options) {
                 if ($app['dbs.default'] === $name) {
                     // we use shortcuts here in case the default has been overridden
@@ -85,7 +86,7 @@ class DoctrineServiceProvider implements ServiceProviderInterface
         $app['dbs.config'] = $app->share(function ($app) {
             $app['dbs.options.initializer']();
 
-            $configs = new \Pimple();
+            $configs = new Container();
             foreach ($app['dbs.options'] as $name => $options) {
                 $configs[$name] = new Configuration();
 
@@ -100,7 +101,7 @@ class DoctrineServiceProvider implements ServiceProviderInterface
         $app['dbs.event_manager'] = $app->share(function ($app) {
             $app['dbs.options.initializer']();
 
-            $managers = new \Pimple();
+            $managers = new Container();
             foreach ($app['dbs.options'] as $name => $options) {
                 $managers[$name] = new EventManager();
             }