+++ src/Silex/Provider/ServiceControllerServiceProvider.php
@@ -11,13 +11,14 @@
 
 namespace Silex\Provider;
 
+use Pimple\Container;
+use Pimple\ServiceProviderInterface;
 use Silex\Application;
-use Silex\ServiceProviderInterface;
 use Silex\ServiceControllerResolver;
 
 class ServiceControllerServiceProvider implements ServiceProviderInterface
 {
-    public function register(Application $app)
+    public function register(Container $app)
     {
         $app['resolver'] = $app->share($app->extend('resolver', function ($resolver, $app) {
             return new ServiceControllerResolver($resolver, $app);