+++ tests/Silex/Tests/ServiceControllerResolverRouterTest.php
@@ -22,7 +22,7 @@ class ServiceControllerResolverRouterTest extends RouterTest
 {
     public function testServiceNameControllerSyntax()
     {
-        $app = new Application();
+        $app = $this->createApplication();
 
         $app['service_name'] = function () {
             return new MyController;
@@ -33,12 +33,10 @@ class ServiceControllerResolverRouterTest extends RouterTest
         $this->checkRouteResponse($app, '/bar', 'bar');
     }
 
-    protected function checkRouteResponse($app, $path, $expectedContent, $method = 'get', $message = null)
+    protected function createApplication()
     {
+        $app = parent::createApplication();
         $app->register(new ServiceControllerServiceProvider());
-
-        $request = Request::create($path, $method);
-        $response = $app->handle($request);
-        $this->assertEquals($expectedContent, $response->getContent(), $message);
+        return $app;
     }
 }